Convert parts of `AstData` into visitors
